integrated camera handler into attachment component
